refactoring openAI targets
