+++ pyrit/chat/__init__.py
@@ -3,6 +3,5 @@
 
 from pyrit.chat.aml_online_endpoint_chat import AMLOnlineEndpointChat
 from pyrit.chat.hugging_face_chat import HuggingFaceChat
-from pyrit.chat.openai_chat import OpenAIChat
 
-__all__ = ["AzureOpenAIChat", "AMLOnlineEndpointChat", "HuggingFaceChat", "OpenAIChat"]
+__all__ = ["AMLOnlineEndpointChat", "HuggingFaceChat"]
