+++ pyrit/prompt_target/__init__.py
@@ -4,7 +4,7 @@
 from pyrit.prompt_target.prompt_target import PromptTarget
 from pyrit.prompt_target.prompt_chat_target import PromptChatTarget
 from pyrit.prompt_target.azure_blob_storage_target import AzureBlobStorageTarget
-from pyrit.prompt_target.azure_openai_chat_target import AzureOpenAIChatTarget
+from pyrit.prompt_target.openai_chat_target import AzureOpenAIChatTarget, OpenAIChatTarget
 from pyrit.prompt_target.gandalf_target import GandalfTarget
 from pyrit.prompt_target.text_target import TextTarget
 
@@ -13,6 +13,7 @@ __all__ = [
     "AzureBlobStorageTarget",
     "AzureOpenAIChatTarget",
     "GandalfTarget",
+    "OpenAIChatTarget",
     "PromptChatTarget",
     "PromptTarget",
     "TextTarget",
