+++ tests/test_openai_chat_target.py
@@ -11,10 +11,8 @@ import pytest
 from openai.types.chat import ChatCompletion, ChatCompletionMessage
 from openai.types.chat.chat_completion import Choice
 
-from pyrit.prompt_target import AzureOpenAIChatTarget
+from pyrit.prompt_target import AzureOpenAIChatTarget, OpenAIChatTarget
 from pyrit.models import ChatMessage
-
-_loop = asyncio.get_event_loop()
 
 
 @pytest.fixture
@@ -36,12 +34,20 @@ def openai_mock_return() -> ChatCompletion:
 
 
 @pytest.fixture
-def chat_engine() -> AzureOpenAIChatTarget:
+def azure_chat_engine() -> AzureOpenAIChatTarget:
     return AzureOpenAIChatTarget(
         deployment_name="gpt-4",
         endpoint="https://mock.azure.com/",
         api_key="mock-api-key",
         api_version="some_version",
+    )
+
+@pytest.fixture
+def openai_chat_engine() -> OpenAIChatTarget:
+    return OpenAIChatTarget(
+        deployment_name="gpt-4",
+        endpoint="https://mock.azure.com/",
+        api_key="mock-api-key",
     )
 
 
@@ -74,21 +80,35 @@ class MockChatCompletionsAsync(AbstractAsyncContextManager):
     "openai.resources.chat.AsyncCompletions.create",
     new_callable=lambda: MockChatCompletionsAsync(),
 )
-def test_complete_chat_async_return(mock_chat_create: AsyncMock, chat_engine: AzureOpenAIChatTarget):
-    ret = _loop.run_until_complete(
-        chat_engine.complete_chat_async(messages=[ChatMessage(role="user", content="hello")])
-    )
-    assert ret == "hi"
 
-
-def test_complete_chat_return(openai_mock_return: ChatCompletion, chat_engine: AzureOpenAIChatTarget):
+@pytest.mark.asyncio
+async def test_azure_complete_chat_async_return(openai_mock_return: ChatCompletion, azure_chat_engine: AzureOpenAIChatTarget):
     with patch("openai.resources.chat.Completions.create") as mock_create:
         mock_create.return_value = openai_mock_return
-        ret = chat_engine.complete_chat(messages=[ChatMessage(role="user", content="hello")])
+        ret = await azure_chat_engine.complete_chat_async(messages=[ChatMessage(role="user", content="hello")])
+        assert ret == "hi"
+
+@pytest.mark.asyncio
+async def test_openai_complete_chat_async_return(openai_mock_return: ChatCompletion, openai_chat_engine: OpenAIChatTarget):
+    with patch("openai.resources.chat.AsyncCompletions.create", new_callable=AsyncMock) as mock_create:
+        mock_create.return_value = openai_mock_return
+        ret = await openai_chat_engine.complete_chat_async(messages=[ChatMessage(role="user", content="hello")])
+        assert ret == "hi"
+
+def test_azure_complete_chat_return(openai_mock_return: ChatCompletion, azure_chat_engine: AzureOpenAIChatTarget):
+    with patch("openai.resources.chat.Completions.create") as mock_create:
+        mock_create.return_value = openai_mock_return
+        ret = azure_chat_engine.complete_chat(messages=[ChatMessage(role="user", content="hello")])
+        assert ret == "hi"
+
+def test_openai_complete_chat_return(openai_mock_return: ChatCompletion, openai_chat_engine: OpenAIChatTarget):
+    with patch("openai.resources.chat.Completions.create") as mock_create:
+        mock_create.return_value = openai_mock_return
+        ret = openai_chat_engine.complete_chat(messages=[ChatMessage(role="user", content="hello")])
         assert ret == "hi"
 
 
-def test_invalid_key_raises():
+def test_azure_invalid_key_raises():
     os.environ[AzureOpenAIChatTarget.API_KEY_ENVIRONMENT_VARIABLE] = ""
     with pytest.raises(ValueError):
         AzureOpenAIChatTarget(
@@ -98,14 +118,28 @@ def test_invalid_key_raises():
             api_version="some_version",
         )
 
+def test_openai_invalid_key_raises():
+    os.environ[OpenAIChatTarget.API_KEY_ENVIRONMENT_VARIABLE] = ""
+    with pytest.raises(ValueError):
+        OpenAIChatTarget(
+            deployment_name="gpt-4",
+            endpoint="https://mock.azure.com/",
+            api_key="",
+        )
 
-def test_initialization_with_no_deployment_raises():
+
+def test_azure_initialization_with_no_deployment_raises():
     os.environ[AzureOpenAIChatTarget.DEPLOYMENT_ENVIRONMENT_VARIABLE] = ""
     with pytest.raises(ValueError):
         AzureOpenAIChatTarget()
 
+def test_openai_initialization_with_no_deployment_raises():
+    os.environ[OpenAIChatTarget.DEPLOYMENT_ENVIRONMENT_VARIABLE] = ""
+    with pytest.raises(ValueError):
+        OpenAIChatTarget()
 
-def test_invalid_endpoint_raises():
+
+def test_azure_invalid_endpoint_raises():
     os.environ[AzureOpenAIChatTarget.ENDPOINT_URI_ENVIRONMENT_VARIABLE] = ""
     with pytest.raises(ValueError):
         AzureOpenAIChatTarget(
@@ -113,4 +147,13 @@ def test_invalid_endpoint_raises():
             endpoint="",
             api_key="xxxxx",
             api_version="some_version",
+        )
+
+def test_openai_invalid_endpoint_raises():
+    os.environ[OpenAIChatTarget.ENDPOINT_URI_ENVIRONMENT_VARIABLE] = ""
+    with pytest.raises(ValueError):
+        OpenAIChatTarget(
+            deployment_name="gpt-4",
+            endpoint="",
+            api_key="xxxxx",
         )
